# Record attributes based on slog.Record