--- a/log/internal/go.mod
+++ b/log/internal/go.mod
@@ -13,7 +13,9 @@ require (
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/go-logr/stdr v1.2.2 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
+	go.opentelemetry.io/otel/metric v1.21.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )