--- a/log/internal/slog_test.go
+++ b/log/internal/slog_test.go
@@ -5,7 +5,6 @@ package internal
 
 import (
 	"testing"
-	"time"
 
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/log"
@@ -20,16 +19,17 @@ func TestSlogHandler(t *testing.T) {
 
 	l.InfoContext(ctx, testBody, "string", testString)
 
-	want := log.Record{
-		Body:     testBody,
-		Severity: log.SeverityInfo,
-		Attributes: []attribute.KeyValue{
-			attribute.String("string", testString),
-		},
-	}
+	want := log.Record{}
+	want.SetBody(testBody)
+	want.SetSeverity(log.SeverityInfo)
+	want.AddAttributes(attribute.String("string", testString))
 
-	assert.NotZero(t, spy.Record.Timestamp, "should set a timestamp")
-	spy.Record.Timestamp = time.Time{}
-	assert.Equal(t, want, spy.Record)
-	assert.Equal(t, ctx, spy.Context)
+	assert.Equal(t, testBody, spy.Record.Body())
+	assert.Equal(t, log.SeverityInfo, spy.Record.Severity())
+	assert.Equal(t, 1, spy.Record.AttributesLen())
+	spy.Record.WalkAttributes(func(kv attribute.KeyValue) bool {
+		assert.Equal(t, "string", string(kv.Key))
+		assert.Equal(t, testString, kv.Value.AsString())
+		return true
+	})
 }