--- a/log/internal/writer_logger_test.go
+++ b/log/internal/writer_logger_test.go
@@ -17,17 +17,16 @@ func TestWriterLogger(t *testing.T) {
 	sb := &strings.Builder{}
 	l := &writerLogger{w: sb}
 
-	r := log.Record{
-		Timestamp: testTimestamp,
-		Severity:  testSeverity,
-		Body:      testBody,
-		Attributes: []attribute.KeyValue{
-			attribute.String("string", testString),
-			attribute.Float64("float", testFloat),
-			attribute.Int("int", testInt),
-			attribute.Bool("bool", testBool),
-		},
-	}
+	r := log.Record{}
+	r.SetTimestamp(testTimestamp)
+	r.SetSeverity(testSeverity)
+	r.SetBody(testBody)
+	r.AddAttributes(
+		attribute.String("string", testString),
+		attribute.Float64("float", testFloat),
+		attribute.Int("int", testInt),
+		attribute.Bool("bool", testBool),
+	)
 	l.Emit(ctx, r)
 
 	want := "timestamp=595728000 severity=9 body=log message string=7e3b3b2aaeff56a7108fe11e154200dd/7819479873059528190 float=1.2345 int=32768 bool=true traced=true\n"