--- a/log/record.go
+++ b/log/record.go
@@ -184,9 +184,6 @@ func countInvalidAttrs(as []attribute.KeyValue) int {
 //
 // This is a copy from https://pkg.go.dev/slices as it is not available in Go 1.20.
 func sliceGrow[S ~[]E, E any](s S, n int) S {
-	if n < 0 {
-		panic("cannot be negative")
-	}
 	if n -= cap(s) - len(s); n > 0 {
 		s = append(s[:cap(s)], make([]E, n)...)[:len(s)]
 	}