… log/go.sum
… log/internal/go.sum
… log/severity_test.go